Moved the testPane commands back to testPane.ts, and left only the more complex clicast stuff in vcastUtilities.
